fixed stats and clock and for ce[f],zio[f] instrumented SIGINT stopping and fixed pause, stop, exit

+++ fs2/main/src/main/scala/dump.scala
@@ -36,7 +36,7 @@ import `Π-loop`.*
 
 package object `Π-dump`:
 
-  type -[F[_]] = Queue[F, Option[(Long, ((Long, Long), (Long, Double)), (String, String), (Double, Double), ((String, (String, String)), (String, (String, String))))]]
+  type -[F[_]] = Queue[F, Option[((Long, Double), ((Long, Long), Long), (String, String), (Double, Double), ((String, (String, String)), (String, (String, String))))]]
 
 
   final class πdump[F[_]: Concurrent]:
