+++ ziof/main/src/main/scala/dump.scala
@@ -35,7 +35,7 @@ import `Π-loop`.*
 
 package object `Π-dump`:
 
-  type - = Queue[Option[(Long, ((Long, Long), (Long, Double)), (String, String), (Double, Double), ((String, (String, String)), (String, (String, String))))]]
+  type - = Queue[Option[((Long, Double), ((Long, Long), Long), (String, String), (Double, Double), ((String, (String, String)), (String, (String, String))))]]
 
 
   def dump(using % : %, ! : !, - : -): UIO[Unit] =
